generator/requirements.txt
tornado
pillow
pandas
redis
numpy